add test demonstrating problem with linknodes
Summary:
Adds a mononoke integration test demonstrating the problem with linknodes and
commit cloud commits, where linknodes may refer to draft commits you don't have
locally, because the filenode was first created with a draft commit that was
pushed to commit cloud.

This results in Mercurial having to do expensive adjustlinkrev operations
on public commits.

Reviewed By: StanislavGlebik

Differential Revision: D13026613

fbshipit-source-id: c2b8e5369a1ba69a175a49b57a70c3e35f92567f