don't store filenodes for draft commits in the filenodes table
Summary:
When receiving and infinitepush bundle, don't store the filenodes for the
commit in the filenodes table.

When a client pulls these commits, we will reconstruct the filenode info from
the blobstore data.  However, they will receive null linknodes, and will need
to use adjustlinknode to compute the correct linknode.

Reviewed By: StanislavGlebik

Differential Revision: D13467110

fbshipit-source-id: 739b06f30a530a159352ffbf612d136c9c831aeb