getbundle request: bundlecaps param should be a set
Summary:
See representation from Mercurial.

```
gboptsmap = {
    "heads": "nodes",
    "bookmarks": "boolean",
    "common": "nodes",
    "obsmarkers": "boolean",
    "phases": "boolean",
    "bundlecaps": "scsv",
    "listkeys": "csv",
    "cg": "boolean",
    "cbattempted": "boolean",
}
```

Some logic might need to check what caps client supports. So convenient lookup is needed.

For example for phases, when we choose to transport phases via a separate bundle2 section called "phase-heads", we need to check that it is a supported on the client side.

Reviewed By: StanislavGlebik

Differential Revision: D13519888

fbshipit-source-id: 6471f2a5057cc3bb2dd72d9318a6f253a447b758