commitcloud: don't updateonmove if the new commit is public
Summary:
The pullcreatemarkers extension interacts badly with commit cloud sync's
updateonmove.  If the current commit has been landed, the next cloud sync
will follow the marker to the landed commit.  This is not usually what the user
wants, and slows down usage as they have to wait for the update to finish
before they can continue working.

Reviewed By: liubov-dmitrieva

Differential Revision: D13599477

fbshipit-source-id: f25d50d9dcb023894f2459e632fbd5ff4d172dd0