mononoke: make sure we don't have stackoverflows in reachability_query
Summary:
reachability_query uses recursion so we can run out of stack. However getting
rid of recursion is easy because it was already implemented for `lca_hint`
method.

I also used this diff as an opportunity to rename `src_hash` and `dst_hash` to
`maybe_descendant_hash` and `maybe_ancestor_hash` respectively to make code
clearer.

Reviewed By: lukaspiatkowski

Differential Revision: D13650422

fbshipit-source-id: 0e52ae592992208a03691b1a5c24021a4fb94313