hgsubversion: dont circumvent revmap abstraction in updatemeta
Summary:
In D13433558 we moved lastpulled into the sqlite database for sqlite
backed revmaps. It turns out the update/rebuildmeta code circumvents the revmap
abstraction and attempts to read the lastpulled file directly from disk. In a
sql backed world, this attempt fails and it then rebuilds the entire revmap
which is very slow.

There was a comment stating the code was intentionally not reading from the
revmap, but I don't believe it applies anymore, and the tests pass fine with
this new change.

Reviewed By: singhsrb

Differential Revision: D13662697

fbshipit-source-id: 2db8f346d89053604d34fbda8f531f688cf71210