mononoke: remove CachingChangesets
Summary:
There is no much point in keeping since we have skiplist which should solve the
same problems in a better way.

The only way where CachingChangesets maybe useful is when many users fetch a
lot of commits simultaneously. It may happen when we merge a new big repository.
However current implementation of CachingChangesets won't help with it since we
do not update its indexes.

Reviewed By: lukaspiatkowski

Differential Revision: D13695201

fbshipit-source-id: 2a4600eccf8224453ca13047e5a2ef3a0af650e3