lfs: fix copy metadata in pack files during lfs moves
Summary:
We should not be attaching copy headers to lfs metadata blobs since
they store the data separately. All this logic was a little overly complicated.
The file text we receive should be passed through unmodified in all cases except
when storing in the loose file format, which needs to strip the copy metadata
from the blob.

Reviewed By: ikostia

Differential Revision: D13703708

fbshipit-source-id: d7e4eb703599d2355009efe070833013d1fcd4fc