Reject empty path or type in repo alias config
Summary:
As of D13737806, for most config options, an option with an empty value is treated as if it is unset. The `type` and `path` options in repository sections behave differently: an empty value is different from no value. Make `type` and `path` consistent with other config options.

This diff should not change behavior given valid config files.

Reviewed By: chadaustin

Differential Revision: D13737986

fbshipit-source-id: f1d2044229d828cc3a04ce46acc31e90bee76c39