refactor some of the checkout handling in "eden doctor"
Summary:
This adds a new `CheckoutInfo` class in the `eden doctor` code to consolidates
some information about a particular checkout.  This should make it easier to
pass around this single object to various check functions, instead of several
separate arguments.

I did make a few functional changes as part of this refactoring:
- All checkouts are now checked in order sorted by their mount path.
  Previously non-mounted checkouts were checked before running checkouts.
- The code now reports additional errors about checkouts if they are running
  but not listed in the on-disk configuration, or if their running state
  differs from the on-disk configuration in some cases.

Reviewed By: strager

Differential Revision: D13526754

fbshipit-source-id: 81802f96a3a3c9f2aaad9a4468fb666ed8b2e47e