Properly support booleans in configs
Summary:
The service.experimental_systemd config option is implemented hackily: the value is parsed as a boolean, then converted to a string (with `str`), then compared with `"True"`. Remove this ugliness by not converting all config options to strings in EdenConfigParser, and adding a type-safe `get_bool` accessor.

Aside from disallowing `experimental_systemd = "True"` in configs and changing the output of `eden config`, this diff should not change behavior.

Reviewed By: simpkins

Differential Revision: D13666331

fbshipit-source-id: 65b7d233c43db08bc7e0288f248041ca4a18a5fd