Implement UnionType for Revision in Thrift parameters
Summary: This is to take the guesswork out of revisions. MononokeRevision now states explicitely whether the value is a hash or bookmark. To roll this out I will prepare an updated scmquery-proxy version, test everything on shadow. Then use the sitevar to disable mononoke usage for a short time while both apisever and scmquery-proxy are rolled out simultaneously

Reviewed By: StanislavGlebik

Differential Revision: D13817040

fbshipit-source-id: d5eee7cf9ac972fb1313a10a17bf60c4545054af