move to the same hg tests runner that is used by hg
Summary:
This is the first step, I just copied the files from mercurial to mononoke/tests/integration/third_party and made it work.

Further unification is required to reuse the same files.

It renders progress bars and other fancy things and matches our mercurial test framework.

Also it allows to use embedded python code!

Some things are not supported: like aliases, so changes in the tests are required.

Some  tests used urandom and cat in a wrong way, I had to fix it as well.

Reviewed By: lukaspiatkowski

Differential Revision: D13871531

fbshipit-source-id: c723bb18c2233639bb36bc7964e57baddff4c0b9