make test infra files to be symlinks to the mercurial ones.
Summary:
I already made them the same (but copy) in a different diff. As we discussed at
war room, symlinks is a better solution.

Also removed unused import and some functions that were copied from mercurial.
This is no longer needed as we now share tinit.sh with general purpose
functions.

The only mononoke specific file left is dummyssh.py

I think later we could pack them in a separate buck target or something and share in a better way that symlinks but for now it is the easiest solution.

Reviewed By: DurhamG

Differential Revision: D13881960

fbshipit-source-id: 36f425d6f0ddbae2c9d083de35d2779669dc01e7