Add API server endpoint to serve HG file content
Summary: Add a `/getfile/` endpoint that takes a filenode and returns Mercurial file content. This is similar to the `/blob/` endpoint, except the returned content can also include copy information in the case that the file was copied. This matches the exact format in which Mercurial stores file content, so the resulting output can be directly written to disk.

Reviewed By: chadaustin

Differential Revision: D13876096

fbshipit-source-id: 0417cb1f993c4c93687e364b1647490c3a9040bc