update some of the doctor tests to use EdenCheckout
Summary:
Update the `create_test_mount()` helper function to return an `EdenCheckout`
object instead of simply the mount path as a string.  This will make it easier
to write some upcoming tests that need a full `EdenCheckout` object.

Since `EdenCheckout` stores the mount path as a `pathlib.Path` instead of as a
plain `str`, this also updates a few more locations in the test code to use
`Path`.

Reviewed By: chadaustin

Differential Revision: D13928666

fbshipit-source-id: 585f27c344e2d0283bf5f8562e17ac3fd356c539