refactor doctor checks for repairing the .hg/dirstate
Summary:
Refactor the checks that verify the contents of `.hg/dirstate`, in preparation
for teaching `eden doctor` how to repair the other required files in the
`.hg/` directory.

This creates a new `HgChecker` class, which I plan to use in the future for
adding checks for each specific file that needs to be checked.  This also
consolidates all issues with the .hg directory into a new `HgDirectoryError`
class.  This is mainly to keep the error reporting simpler: if the entire .hg
directory is missing we will report this as a single problem rather than a
separate problem for each file in the .hg directory that we will recreate.

Reviewed By: chadaustin

Differential Revision: D13928667

fbshipit-source-id: fb78f014d44e85b9ac07048cbe641e9e8be4ed4c